Organized header and source files into separate folders
