+++ ini-rw/inc/IniSetting.hpp
