+++ ini-rw/inc/algorithms/Search.hpp
