+++ ini-rw/inc/entities/IniEntity.hpp
