+++ ini-rw/inc/entities/IniNewLine.hpp
