+++ ini-rw/inc/enums/IniEntityType.hpp
