+++ ini-rw/src/entities/IniNewLine.cpp
